Rename Components to Vector
The code sounds better: `Vector([Trend(1), Trend(2)])` reads as "vector
trend" which is what it actually is.
